feat(vs2022_third): C4244 fix all conversions in StressLineSectionSettingDlg.cpp — 1 file
Fix 13 C4244 warnings:
- Line 1068: Add casts for CRect constructor (4 conversions)
- Lines 2655, 2658: Add casts for MoveTo/LineTo arguments (4 conversions)
- Line 2817: Add cast for AddPoint arguments (2 conversions)
- Lines 2968-2970: Add casts for double→LONG assignments (3 conversions)

No behavior changes, only explicit type conversions.

+++ SRC/NanoSpec/StressLineSectionSettingDlg.cpp
@@ -1065,7 +1065,7 @@ void CStressLineSectionSettingDlg::OnPaint()
 		m_dDrawEndStagePosY = StageRectBottom;
 	}
 
-	pDC->Rectangle(CRect(m_dDrawStartStagePosX, m_dDrawStartStagePosY, m_dDrawEndStagePosX, m_dDrawEndStagePosY));
+	pDC->Rectangle(CRect(static_cast<int>(m_dDrawStartStagePosX), static_cast<int>(m_dDrawStartStagePosY), static_cast<int>(m_dDrawEndStagePosX), static_cast<int>(m_dDrawEndStagePosY)));
 	pDC->SelectObject(m_oldPen);
 	pDC->SelectObject(m_oldBrush);
 
@@ -2814,7 +2814,7 @@ void CStressLineSectionSettingDlg::SamplePointGraph_DataSet()
 				break;
 			}
 
-			m_SamplePointGraph.AddPoint(dPointX, dPointY);
+			m_SamplePointGraph.AddPoint(static_cast<long>(dPointX), static_cast<long>(dPointY));
 // 2009.09.07 bagus stress Œ´“_ˆÊ’u‘Î‰ž --}--
 		}
 	}
